+++ example/demo1.f90
@@ -1,13 +1,14 @@
 !> Backward auto diff
 program main
-    use auto_diff, only: tree_t, operator(*), operator(+), exp, rk
+
+    use auto_diff, only: tree_t, operator(*), operator(+), exp, rk, assignment(=)
     implicit none
     type(tree_t) :: a, b, c
     type(tree_t) :: y
 
-    call a%constructor(value=2.0_rk)
-    call b%constructor(value=1.0_rk)
-    call c%constructor(value=0.0_rk)
+    a = 2.0_rk
+    b = 1.0_rk
+    c = 0.0_rk
 
     print *, "demo1: y = (a + b*b)*exp(c)"
 
@@ -22,8 +23,8 @@ program main
 
     ! - - -
     
-    call a%constructor(value=2.0_rk)
-    call b%constructor(value=1.0_rk)
+    a = 2.0_rk
+    b = 1.0_rk
     
     print *, ""
     print *, "demo2: y = (a + b)*(b + 1.0)"