+++ src/ad_usr_func.f90
@@ -1,7 +1,7 @@
 module ad_usr_func
 
     use ad_kinds, only: rk
-    use ad_types, only: tree_t
+    use ad_types, only: tree_t, assignment(=)
     implicit none
     private
     
@@ -13,8 +13,7 @@ contains
         type(tree_t), intent(in) :: t1
         type(tree_t) :: t
         
-        allocate(t%node)
-        t%node%value = 1.0_rk/(1.0_rk + exp(-t1%node%value))
+        t = 1.0_rk/(1.0_rk + exp(-t1%node%value))
         
         t%node%left => t1%node
         t%node%left_grad = t%node%value*(1.0_rk - t%node%value)