+++ test/test_func.f90
@@ -1,8 +1,9 @@
 module test_func
+
     use testdrive, only: new_unittest, unittest_type, error_type, check
     use auto_diff, only: abs, exp, sqrt, sin, cos, tan, log, log10
-    use auto_diff, only: max, min
-    use auto_diff, only: tree_t, rk
+    use auto_diff, only: max, min, sigmoid
+    use auto_diff, only: tree_t, rk, assignment(=)
     implicit none
     private
 
@@ -25,7 +26,8 @@ contains
                     new_unittest("func log valid", test_log_valid), &
                     new_unittest("func log10 valid", test_log10_valid), &
                     new_unittest("func max valid", test_max_valid), &
-                    new_unittest("func min valid", test_min_valid) &
+                    new_unittest("func min valid", test_min_valid), &
+                    new_unittest("func sigmoid valid", test_sigmoid_valid) &
                     ]
 
     end subroutine collect_suite_func
@@ -36,11 +38,11 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = abs(a)
         call b%backward()
         
-        call check(error, b%get_value(), 2.0_rk)
+        call check(error, b%get_value(), 2.0_rk); if (allocated(error)) return
         call check(error, a%get_grad(), 1.0_rk)
         
     end subroutine test_abs_valid
@@ -51,11 +53,11 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = sqrt(a)
         call b%backward()
         
-        call check(error, b%get_value(), sqrt(2.0_rk))
+        call check(error, b%get_value(), sqrt(2.0_rk)); if (allocated(error)) return
         call check(error, a%get_grad(), 1.0_rk/(2.0_rk*sqrt(2.0_rk)))
         
     end subroutine test_sqrt_valid
@@ -66,11 +68,11 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = exp(a)
         call b%backward()
         
-        call check(error, b%get_value(), exp(2.0_rk))
+        call check(error, b%get_value(), exp(2.0_rk)); if (allocated(error)) return
         call check(error, a%get_grad(), exp(2.0_rk))
         
     end subroutine test_exp_valid
@@ -81,11 +83,11 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = sin(a)
         call b%backward()
         
-        call check(error, b%get_value(), sin(2.0_rk))
+        call check(error, b%get_value(), sin(2.0_rk)); if (allocated(error)) return
         call check(error, a%get_grad(), cos(2.0_rk))
         
     end subroutine test_sin_valid
@@ -96,11 +98,11 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = cos(a)
         call b%backward()
         
-        call check(error, b%get_value(), cos(2.0_rk))
+        call check(error, b%get_value(), cos(2.0_rk)); if (allocated(error)) return
         call check(error, a%get_grad(), -sin(2.0_rk))
         
     end subroutine test_cos_valid
@@ -111,11 +113,11 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = tan(a)
         call b%backward()
         
-        call check(error, b%get_value(), tan(2.0_rk))
+        call check(error, b%get_value(), tan(2.0_rk)); if (allocated(error)) return
         call check(error, a%get_grad(), 1.0_rk/(cos(2.0_rk)**2))
         
     end subroutine test_tan_valid
@@ -126,11 +128,11 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = log(a)
         call b%backward()
         
-        call check(error, b%get_value(), log(2.0_rk))
+        call check(error, b%get_value(), log(2.0_rk)); if (allocated(error)) return
         call check(error, a%get_grad(), 1.0_rk/2.0_rk)
         
     end subroutine test_log_valid
@@ -141,11 +143,11 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = log10(a)
         call b%backward()
         
-        call check(error, b%get_value(), log10(2.0_rk))
+        call check(error, b%get_value(), log10(2.0_rk)); if (allocated(error)) return
         call check(error, a%get_grad(), 1.0_rk/log(10.0_rk)/2.0_rk)
         
     end subroutine test_log10_valid
@@ -156,11 +158,11 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = max(a, a)
         call b%backward()
         
-        call check(error, b%get_value(), 2.0_rk)
+        call check(error, b%get_value(), 2.0_rk); if (allocated(error)) return
         call check(error, a%get_grad(), 1.0_rk)
         
     end subroutine test_max_valid
@@ -171,19 +173,18 @@ contains
         type(tree_t) :: a
         type(tree_t) :: b
         
-        call a%constructor(value=2.0_rk)
+        a = 2.0_rk
         b = min(a, a)
         call b%backward()
         
-        call check(error, b%get_value(), 2.0_rk)
+        call check(error, b%get_value(), 2.0_rk); if (allocated(error)) return
         call check(error, a%get_grad(), 1.0_rk)
         
-        ! call b%destructor()
-        call a%constructor(value=1.0_rk)
+        a = 1.0_rk
         b = min(a, 2.0_rk)
         call b%backward()
         
-        call check(error, b%get_value(), 1.0_rk)
+        call check(error, b%get_value(), 1.0_rk); if (allocated(error)) return
         call check(error, a%get_grad(), 1.0_rk)
         
         !> Not implemented yet
@@ -195,5 +196,20 @@ contains
         ! call check(error, a%get_grad(), 1.0_rk)
         
     end subroutine test_min_valid
+    
+    subroutine test_sigmoid_valid(error)
+        type(error_type), allocatable, intent(out) :: error
+        
+        type(tree_t) :: a
+        type(tree_t) :: b
+        
+        a = 2.0_rk
+        b = sigmoid(a)
+        call b%backward()
+        
+        call check(error, b%get_value(), 1.0_rk/(1.0_rk+exp(-2.0_rk))); if (allocated(error)) return
+        call check(error, a%get_grad(), 0.10499358540350662_rk)
+        
+    end subroutine test_sigmoid_valid
 
 end module test_func