+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 - [x] Change the data structure, hide pointer
 - [x] Refactor the code
 - [x] Support arrays
+- [x] Add a benchmark
+- [x] Final procedure of `tree_t`: Not sure if this will ocurr memory leak
+- [x] Support `assignment(=)` and `tree_t` constructor.
 
 ## TO DO
 
-- [ ] Better support for arrays
-- [ ] Destructor, avoid memory leaks
+- [ ] Better support for arrays (?)
+- [ ] Destructor, avoid memory leaks (?)
+- [ ] Support `integer`